+++ EatPic-iOS/Sources/Components/Calendar/View/CalendarScrollView.swift
@@ -27,106 +27,6 @@ struct MonthVisiblePreferenceKey: PreferenceKey {
     }
 }
 
-/// (연, 월) 조합을 표현하는 경량 키 타입입니다.
-/// - Important: `Hashable`을 채택해 Set/Dictionary의 키로 사용합니다.
-struct YearMonth: Hashable {
-    let year: Int
-    let month: Int
-}
-
-/// 서버 공통 응답 래퍼입니다.
-/// - Parameters:
-///   - T: `result` 페이로드 타입(배열/단일 모두 지원).
-/// - Tip: 다양한 API에서 **공통 래퍼 재사용**을 위해 제네릭으로 설계되어 있습니다.
-struct APIResponse<T: Codable>: Codable {
-    let isSuccess: Bool
-    let code: String
-    let message: String
-    let result: T
-}
-
-/// 캘린더 한 항목에 대한 응답 DTO입니다.
-/// - Note: `date`는 문자열로 내려오므로, 파싱이 필요하면 별도 매퍼에서 처리하세요.
-struct CalendarResponse: Codable {
-    let date: String
-    let imageUrl: String
-    let cardId: Int
-}
-
-/// 캘린더 무한 스크롤과 월별 데이터 로딩을 관리하는 뷰모델입니다.
-/// - Responsibilities:
-///   - `visibleYM` 변화에 맞춰 **월별 데이터 페치**
-///   - **중복 호출 방지**: 이미 로드한 월은 `loaded` 캐시로 차단
-///   - 네트워크/디코딩은 백그라운드에서 수행, UI 반영만 MainActor에서 처리
-@Observable
-final class CalendarScrollViewModel {
-    /// 홈 모듈용 `MoyaProvider`
-    private let homeProvider: MoyaProvider<HomeTargetType>
-    /// 이미 성공적으로 로드한 (연, 월) 키 캐시. 중복 호출 방지 용도입니다.
-    private var loaded: Set<YearMonth> = []
-    
-    /// - Parameter container: DI 컨테이너. 여기서 필요한 Provider를 주입받습니다.
-    init(container: DIContainer) {
-        self.homeProvider = container.apiProviderStore.home()
-    }
-    
-    /// 특정 연/월의 캘린더 데이터를 요청합니다.
-    /// - Parameters:
-    ///   - year: 연도
-    ///   - month: 월
-    /// - Important:
-    ///   - **중복 호출 방지**: 이미 로드된 `(year, month)`는 재요청하지 않습니다.
-    ///   - 네트워크/디코딩은 백그라운드, UI 반영만 필요 시 `MainActor`에서 수행하세요.
-    func fetchCalendarData(year: Int, month: Int) async {
-        let key = YearMonth(year: year, month: month)
-        guard !loaded.contains(key) else { return }
-        
-        do {
-            let response = try await homeProvider.requestAsync(
-                .fetchCalendar(year: year, month: month))
-            let data = try JSONDecoder().decode(
-                APIResponse<[CalendarResponse]>.self, from: response.data)
-            
-            guard data.isSuccess else {
-                throw APIError.server(code: data.code, message: data.message)
-            }
-            
-            // 성공 시, 필요한 상태 업데이트를 메인에서 처리.
-            // await MainActor.run { ... }
-            loaded.insert(key)
-            print(data)
-        } catch let error as MoyaError {
-            let msg = readable(error)
-            print("MoyaError:", msg)
-        } catch {
-            print("Decode/Other Error:", error.localizedDescription)
-        }
-    }
-    
-    /// 사용자/로깅을 위한 읽기 쉬운 에러 메시지를 생성합니다.
-    private func readable(_ error: MoyaError) -> String {
-        switch error {
-        case let .statusCode(res): return "서버 오류(\(res.statusCode))"
-        case let .underlying(err, _): return "네트워크 오류: \(err.localizedDescription)"
-        default: return "요청 실패: \(error.localizedDescription)"
-        }
-    }
-}
-
-// MARK: - 에러 보조
-
-/// 서버가 `isSuccess = false`로 내려줄 때를 표현하는 에러 타입입니다.
-enum APIError: LocalizedError {
-    case server(code: String, message: String)
-    
-    var errorDescription: String? {
-        switch self {
-        case let .server(code, message):
-            return "[\(code)] \(message)"
-        }
-    }
-}
-
 /// 달력을 무한 스크롤로 여러 달을 연속적으로 보여주는 뷰입니다.
 /// - Features:
 ///   - **GeometryReader + PreferenceKey**로 각 월의 화면 내 상대 위치를 보고
